issue : 파일 의존성 분리[COMMON-INTF6]
