+++ traffic_control_system/traffic_control_system/.settings/org.eclipse.core.resources.prefs
@@ -23,6 +23,12 @@ encoding//Debug/eddi/comm/lwip-1.4.1/src/netif/subdir_rules.mk=UTF-8
 encoding//Debug/eddi/comm/lwip-1.4.1/src/netif/subdir_vars.mk=UTF-8
 encoding//Debug/eddi/comm/prj_ethernet/source/subdir_rules.mk=UTF-8
 encoding//Debug/eddi/comm/prj_ethernet/source/subdir_vars.mk=UTF-8
+encoding//Debug/eddi/sensor_control/spi_manage_object/source/subdir_rules.mk=UTF-8
+encoding//Debug/eddi/sensor_control/spi_manage_object/source/subdir_vars.mk=UTF-8
+encoding//Debug/eddi/tms570_common_interface/sensor_control/spi_manage_object/source/subdir_rules.mk=UTF-8
+encoding//Debug/eddi/tms570_common_interface/sensor_control/spi_manage_object/source/subdir_vars.mk=UTF-8
+encoding//Debug/eddi/tms570_common_interface/sensor_control/spi_manage_operation/source/subdir_rules.mk=UTF-8
+encoding//Debug/eddi/tms570_common_interface/sensor_control/spi_manage_operation/source/subdir_vars.mk=UTF-8
 encoding//Debug/makefile=UTF-8
 encoding//Debug/objects.mk=UTF-8
 encoding//Debug/source/subdir_rules.mk=UTF-8
