+++ traffic_control_system/traffic_control_system/eddi/tms570_common_interface/sensor_control/spi_manage_object/source/spi_obj_manager.c
@@ -5,23 +5,8 @@
  *      Author: son
  */
 #include "spi_obj_manager.h"
+#include "spi_operation_manager.h"
 #include "stdlib.h"
-
-typedef spiBASE_t *(*_spi_open)(const spi_dev_num, const sensor_dev_name);
-typedef void (*_spi_close)(const spi_dev_num);
-typedef void (*_spi_ioctl)(void);
-typedef void (*_spi_read)(const spiDAT1_t *, uint16_t *, uint16_t);
-typedef void (*_spi_write)(const spiDAT1_t *, uint16_t *, uint16_t);
-
-typedef struct _spi_operations spi_operations;
-struct _spi_operations{
-    //spi_io_ctl 세부사항 미구현
-    _spi_open open;
-    _spi_close close;
-    _spi_ioctl ioctl;
-    _spi_read read;
-    _spi_write write;
-};
 
 typedef struct _spi_register spi_register;
 struct _spi_register{
@@ -54,6 +39,5 @@ boolean spi_open(const spi_dev_num spi_num, const sensor_dev_name dev_name)
 
 void spi_close(const spi_dev_num spi_num)
 {
-    free(spi_dev_t[spi_num]->operations);
     free(spi_dev_t[spi_num]);
 }
