+++ traffic_control_system/traffic_control_system/eddi/tms570_common_interface/sensor_control/spi_manage_operation/source/spi_operation_manager.c
@@ -1,0 +1,10 @@
+/*
+ * spi_manage_operations.c
+ *
+ *  Created on: 2022. 8. 5.
+ *      Author: son
+ */
+#include "spi_operation_manager.h"
+
+//미구현
+
